test temporary stop

+++ src/main/java/com/example/demo/src/video/dto/VideoResDto.java
@@ -10,9 +10,9 @@ public class VideoResDto {
 
     private String title;
     private String desc;
-    private int dur;
+    private double dur;
 
-    public VideoResDto(String title, String desc, int dur) {
+    public VideoResDto(String title, String desc, double dur) {
         this.title = title;
         this.desc = desc;
         this.dur = dur;
